requirements.txt
PyPDF2==3.0.0
PyMuPDF==1.22.5
Pillow==9.2.0
ImageHash==4.3.1
